serpinski mostly working

+++ src/camera.h
@@ -5,6 +5,6 @@ void camera_init();
 void camera_special_keys(int key, int x, int y);
 void camera_special_keys_up(int key, int x, int y);
 void camera_draw();
-void camera_update(int delta);
+void camera_update(double delta);
 
 #endif
